Pass `DataSource` to `QueryBuilder` (#1922)
## Changes
<!-- Summary of your changes that are easy to understand. Add
screenshots when necessary, they're helpful to illustrate the before and
after state -->
### What does this PR do?
**Big Picture:** 
Add support in lakebridge reconciler for columns with special chars e.g.
`Cust#` as at the moment reconciler will fail if such columns are
passed.

**This PR:**
Add `DataSource` to `QueryBuilder` to allow building source aware
queries. this is a small "refactoring" that adds no new logic but to
make next PR smaller

### Caveats/things to watch out for when reviewing:
This is a big change that had the implementation split between 6 PRs to
make reviews manageable. this cannot go to main without the rest of the
PRs. this is PR 5/6

### Linked issues
<!-- DOC: Link issue with a keyword: close, closes, closed, fix, fixes,
fixed, resolve, resolves, resolved. See
https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword
-->

Resolves #1866 

### Functionality

- [ ] added relevant user documentation
- [ ] added new CLI command
- [ ] modified existing command: `databricks labs lakebridge ...`

### Tests
<!-- How is this tested? Please see the checklist below and also
describe any other relevant tests -->

- [ ] manually tested
- [ ] added unit tests
- [ ] added integration tests