Support columns names having special characters in schema compare (#1921)
<!-- REMOVE IRRELEVANT COMMENTS BEFORE CREATING A PULL REQUEST -->
## Changes
<!-- Summary of your changes that are easy to understand. Add
screenshots when necessary, they're helpful to illustrate the before and
after state -->
### What does this PR do?
**Big Picture:** 
Add support in lakebridge reconciler for columns with special chars e.g.
`Cust#` as at the moment reconciler will fail if such columns are
passed.

**This PR:**
Second step is to support the normalized columns names in schema
compare. We need to update all sqlglot calls with normalized column
names in case the user is using special chars.

### Relevant implementation details
* Schema compare uses sqlglot which expects source compatible dialect so
we cannot use ansi everywhere since e.g. snowflake or oracle use other
characters and not backticks to delimit identifiers

### Caveats/things to watch out for when reviewing:
This is a big change that had the implementation split between 6 PRs to
make reviews manageable. this cannot go to main without the rest of the
PRs. this is PR 2/6

### Linked issues
<!-- DOC: Link issue with a keyword: close, closes, closed, fix, fixes,
fixed, resolve, resolves, resolved. See
https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword
-->

Resolves #1866 

### Functionality

- [ ] added relevant user documentation
- [ ] added new CLI command
- [ ] modified existing command: `databricks labs lakebridge ...`

### Tests
<!-- How is this tested? Please see the checklist below and also
describe any other relevant tests -->

- [ ] manually tested
- [x] added unit tests
- [ ] added integration tests